requirements.txt
flask
groq
flask-cors
python-dotenv
supabase
gunicorn
